Release: sane 7.0
Version 7.0 was completely rewritten from scratch. Sane is now based on
a small number of decorators, namely @cmd, @task, @tag, and @depends.
User input is more thoroughly validated, and sane should be overall
friendlier. The new definitions also allow for more predictable
behaviour, be easier to remember, and result in cleaner code.

Sane 7.0 is also magic: explicitly calling a function to run sane is no
longer necessarily required, as isn't specifying command line mode or
not (which, to be quite honest, was confusing and inelegant). Now
calling a @task or @cmd spawns and runs the tree, as one might expect.

The documentation was also completely re-written, and sane now includes
a manual with a full guide-by-example, and a proper reference for all of
the provided decorators, as well as some commentary on the internal
workings of sane.

The new code was written in a self-documenting style and should be
easier to understand.