Fixes: --jobs, and @default diagnosis
The --jobs flag always failed, because a missing star in the
ThreadExecutor submission meant that the args were passed to the task as
a function argument, instead of being unrolled. This meant that tasks
not taking exactly one argument would crash sane. It seems correct, now.

Using @default after @sane.cmd could cause execution to crash without a
useful diagnosis message. This is because I overlooked the situation
where the function already has a __sane__ properties dictionary, but no
type assigned to it yet (thus, a default None value). This is now
accounted for.